v0.5.2: Fix builds from source in .crate file
The crates uploaded for 0.5.0 and 0.5.1 were completely broken
for automatic builds because two things were excluded that shouldn't
have been:  An (empty) potentials/ directory, and the manpage file.

The manpage file is an easy fix.  Leaving behind an empty directory
for potentials/, however, appears to be impossible.  We have to create
a copy of the lammps source tree in our output directory just so we can
add an empty directory at build script runtime.